Make clean and linkify configurable
`clean` and `linkify` are optional parameters to the `render` function.
The caller can provider their own callable for these or they can set
them to `None` to turn them off completely.

PyPI, in particular, needed the ability to turn `linkify` off, because
it's adding extra `rel="nofollow"` attributes to links that it doesn't
have (but maybe should when it switches to using pypa/readme?)
