+++ readme/linkify.py
@@ -1,0 +1,36 @@
+# Copyright 2014 Donald Stufft
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from __future__ import absolute_import, division, print_function
+
+import bleach
+
+
+def linkify(html):
+    def nofollow(attrs, new=False):
+        if attrs["href"].startswith("mailto:"):
+            return attrs
+        attrs["rel"] = "nofollow"
+        return attrs
+
+    # Bleach Linkify makes it easy to modify links, however, we will not be
+    # using it to create additional links.
+    return bleach.linkify(
+        html,
+        callbacks=[
+            lambda attrs, new: attrs if not new else None,
+            nofollow,
+        ],
+        skip_pre=True,
+        parse_email=False,
+    )
